+++ packages/workbox-background-sync/BackgroundSyncPlugin.mjs
