+++ packages/workbox-broadcast-cache-update/_public.mjs
@@ -7,7 +7,7 @@
 */
 
 import {BroadcastCacheUpdate} from './BroadcastCacheUpdate.mjs';
-import {BroadcastCacheUpdatePlugin} from './Plugin.mjs';
+import {BroadcastCacheUpdatePlugin} from './BroadcastCacheUpdatePlugin.mjs';
 import {broadcastUpdate} from './broadcastUpdate.mjs';
 import {responsesAreSame} from './responsesAreSame.mjs';
 import './_version.mjs';
