+++ packages/workbox-cacheable-response/CacheableResponsePlugin.mjs
