+++ packages/workbox-google-analytics/initialize.mjs
@@ -6,7 +6,8 @@
   https://opensource.org/licenses/MIT.
 */
 
-import {BackgroundSyncPlugin} from 'workbox-background-sync/Plugin.mjs';
+import {BackgroundSyncPlugin} from
+  'workbox-background-sync/BackgroundSyncPlugin.mjs';
 import {cacheNames} from 'workbox-core/_private/cacheNames.mjs';
 import {getFriendlyURL} from 'workbox-core/_private/getFriendlyURL.mjs';
 import {logger} from 'workbox-core/_private/logger.mjs';
