+++ test/workbox-cache-expiration/node/test-Plugin.mjs
@@ -12,7 +12,7 @@ import sinon from 'sinon';
 import expectError from '../../../infra/testing/expectError';
 import {devOnly} from '../../../infra/testing/env-it';
 
-import {ExpirationPlugin as Plugin} from '../../../packages/workbox-cache-expiration/Plugin.mjs';
+import {ExpirationPlugin} from '../../../packages/workbox-cache-expiration/ExpirationPlugin.mjs';
 import {CacheExpiration} from '../../../packages/workbox-cache-expiration/CacheExpiration.mjs';
 import {cacheNames} from '../../../packages/workbox-core/_private/cacheNames.mjs';
 import {executeQuotaErrorCallbacks} from '../../../packages/workbox-core/_private/quota.mjs';
@@ -31,13 +31,13 @@ describe(`[workbox-cache-expiration] Plugin`, function() {
   describe(`constructor`, function() {
     devOnly.it(`should throw for no config`, function() {
       return expectError(() => {
-        new Plugin();
+        new ExpirationPlugin();
       }, 'max-entries-or-age-required');
     });
 
     devOnly.it(`should throw for non-number maxEntries`, function() {
       return expectError(() => {
-        new Plugin({
+        new ExpirationPlugin({
           maxEntries: 'Hi',
         });
       }, 'incorrect-type');
@@ -45,28 +45,28 @@ describe(`[workbox-cache-expiration] Plugin`, function() {
 
     devOnly.it(`should throw for non-number maxAgeSeconds`, function() {
       return expectError(() => {
-        new Plugin({
+        new ExpirationPlugin({
           maxAgeSeconds: 'Hi',
         });
       }, 'incorrect-type');
     });
 
     it(`should construct with just maxAgeSeconds`, function() {
-      const plugin = new Plugin({
+      const plugin = new ExpirationPlugin({
         maxAgeSeconds: 10,
       });
       expect(plugin._maxAgeSeconds).to.equal(10);
     });
 
     it(`should construct with just maxEntries`, function() {
-      const plugin = new Plugin({
+      const plugin = new ExpirationPlugin({
         maxEntries: 10,
       });
       expect(plugin._config.maxEntries).to.equal(10);
     });
 
     it(`should register a quota error callback when purgeOnQuotaError is true`, async function() {
-      const plugin = new Plugin({
+      const plugin = new ExpirationPlugin({
         maxEntries: 10,
         purgeOnQuotaError: true,
       });
@@ -78,7 +78,7 @@ describe(`[workbox-cache-expiration] Plugin`, function() {
     });
 
     it(`should not register a quota error callback when purgeOnQuotaError is false`, async function() {
-      const plugin = new Plugin({
+      const plugin = new ExpirationPlugin({
         maxEntries: 10,
         purgeOnQuotaError: false,
       });
@@ -92,7 +92,7 @@ describe(`[workbox-cache-expiration] Plugin`, function() {
 
   describe(`cachedResponseWillBeUsed()`, function() {
     it(`should expose a cachedResponseWillBeUsed() method`, function() {
-      const plugin = new Plugin({maxAgeSeconds: 1});
+      const plugin = new ExpirationPlugin({maxAgeSeconds: 1});
       expect(plugin).to.respondTo('cachedResponseWillBeUsed');
     });
 
@@ -105,7 +105,7 @@ describe(`[workbox-cache-expiration] Plugin`, function() {
       const dateString = new Date().toUTCString();
       const cachedResponse = new Response('', {headers: {date: dateString}});
 
-      const plugin = new Plugin({maxAgeSeconds: 1});
+      const plugin = new ExpirationPlugin({maxAgeSeconds: 1});
 
       const expirationManager = plugin._getCacheExpiration('test-cache');
       sandbox.spy(expirationManager, 'expireEntries');
@@ -125,7 +125,7 @@ describe(`[workbox-cache-expiration] Plugin`, function() {
       // Clock past the expiration of the Data header
       clock.tick(1000 + 1);
 
-      const plugin = new Plugin({maxAgeSeconds: 1});
+      const plugin = new ExpirationPlugin({maxAgeSeconds: 1});
 
       const expirationManager = plugin._getCacheExpiration('test-cache');
       sandbox.spy(expirationManager, 'expireEntries');
@@ -135,7 +135,7 @@ describe(`[workbox-cache-expiration] Plugin`, function() {
     });
 
     it(`should handle a null cachedResponse`, function() {
-      const plugin = new Plugin({maxAgeSeconds: 1});
+      const plugin = new ExpirationPlugin({maxAgeSeconds: 1});
 
       const expirationManager = plugin._getCacheExpiration('test-cache');
       sandbox.spy(expirationManager, 'expireEntries');
@@ -146,13 +146,13 @@ describe(`[workbox-cache-expiration] Plugin`, function() {
 
   describe(`_isResponseDateFresh()`, function() {
     it(`should return true when maxAgeSeconds is not set`, function() {
-      const plugin = new Plugin({maxEntries: 1});
+      const plugin = new ExpirationPlugin({maxEntries: 1});
       const isFresh = plugin._isResponseDateFresh(new Response('Hi'));
       expect(isFresh).to.equal(true);
     });
 
     it(`should return true when there is no Date header`, function() {
-      const plugin = new Plugin({maxAgeSeconds: 1});
+      const plugin = new ExpirationPlugin({maxAgeSeconds: 1});
       const isFresh = plugin._isResponseDateFresh(new Response('Hi', {
         // TODO: Remove this when https://github.com/pinterest/service-workers/issues/72
         // is fixed.
@@ -162,7 +162,7 @@ describe(`[workbox-cache-expiration] Plugin`, function() {
     });
 
     it(`should return true when the Date header is invalid`, function() {
-      const plugin = new Plugin({maxAgeSeconds: 1});
+      const plugin = new ExpirationPlugin({maxAgeSeconds: 1});
       const isFresh = plugin._isResponseDateFresh(new Response('Hi', {
         headers: {date: 'invalid header'},
       }));
@@ -172,7 +172,7 @@ describe(`[workbox-cache-expiration] Plugin`, function() {
 
   describe(`cacheDidUpdate()`, function() {
     it(`should expose a cacheDidUpdate() method`, function() {
-      const plugin = new Plugin({maxAgeSeconds: 1});
+      const plugin = new ExpirationPlugin({maxAgeSeconds: 1});
       expect(plugin).to.respondTo('cacheDidUpdate');
     });
 
@@ -180,7 +180,7 @@ describe(`[workbox-cache-expiration] Plugin`, function() {
       const cacheName = 'test-cache';
       const url = new URL('/test', self.location).toString();
       const request = new Request(url);
-      const plugin = new Plugin({maxAgeSeconds: 10});
+      const plugin = new ExpirationPlugin({maxAgeSeconds: 10});
 
       sandbox.spy(CacheExpiration.prototype, 'updateTimestamp');
       sandbox.spy(CacheExpiration.prototype, 'expireEntries');
@@ -195,7 +195,7 @@ describe(`[workbox-cache-expiration] Plugin`, function() {
 
   describe(`_getCacheExpiration()`, function() {
     it(`should reject when called with the default runtime cache name`, async function() {
-      const plugin = new Plugin({maxAgeSeconds: 1});
+      const plugin = new ExpirationPlugin({maxAgeSeconds: 1});
       await expectError(
           () => plugin._getCacheExpiration(cacheNames.getRuntimeName()),
           'expire-custom-caches-only'
